backend/requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
motor
passlib[bcrypt]
bcrypt==4.0.1
numpy
scikit-learn
python-multipart
requests
python-dotenv
python-jose[cryptography]
gunicorn
dnspython